requirements.txt
opencv-python
matplotlib
PyYAML
numpy
scipy
torch
torchvision
torchaudio

-e .